+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ChangeLog for chef-serverdensity
 ================================
+3.2.3 (20-12-2018)
+------------------
+- Fixed and improved docker daemon plugin configs.
+- Added support for TCP plugin
+
 3.2.2 (15-10-2018)
 ------------------
 - Adds support for Ubuntu Cosmic.